prevent int overflow for over 4GB/s

+++ read-trend.c
@@ -131,15 +131,15 @@ int main(int argc, char *argv[])
     struct timeval start;
     gettimeofday(&start, NULL);
 
-    int read_bytes = 0;
-    int read_count = 0;
+    long read_bytes = 0;
+    long read_count = 0;
     for ( ; ; ) {
         if (has_alarm) {
             has_alarm = 0;
             struct timeval now, elapse;
             gettimeofday(&now, NULL);
             timersub(&now, &start, &elapse);
-            fprintf(stderr, "%ld.%06ld %.3f MB %d\n", elapse.tv_sec, elapse.tv_usec, read_bytes/1024.0/1024.0, read_count);
+            fprintf(stderr, "%ld.%06ld %.3f MB %ld\n", elapse.tv_sec, elapse.tv_usec, read_bytes/1024.0/1024.0, read_count);
             read_bytes = 0;
             read_count = 0;
         }
